+++ frontend/src/features/reviews/index.ts
@@ -1,0 +1,13 @@
+export { default as ReviewItem } from './components/ReviewItem';
+export { default as ReviewList } from './components/ReviewList';
+
+export * from './api/reviews.api';
+
+export * from './hooks/useReviews';
+export * from './hooks/useInfiniteReviews';
+
+export * from './model/reviews.model';
+
+export * from './components/ReviewItem.style';
+
+export { default as formatReviews } from './utils/formatReviews';
