+++ frontend/src/features/reviews/model/reviews.model.ts
@@ -1,0 +1,14 @@
+export interface IResponseReviews {
+  categoryId: number;
+  content: string;
+  createdAt: Date;
+  id: number;
+  isMyReview: number;
+  likes: number;
+  reviewImg: string;
+  stars: number;
+  title: string;
+  userId: number;
+  userName: string;
+  verified: number;
+}
