+++ frontend/src/pages/Join/JoinStep2Nickname.tsx
@@ -1,13 +1,11 @@
 import { useForm } from 'react-hook-form';
 
 import Layout from '@/layout/Layout';
-import JoinTemplate from '@/components/Join/JoinTemplate';
-import { AlertText, Input } from '@/components/common';
+import { JoinTemplate, IUserRegistration, useAuth } from '@/features/auth';
+import { AlertText, Input } from '@/shared/components';
 import { withUnauthenticatedUser } from '@/shared/hocs';
-import { IUserRegistration } from '@/models/user.model';
 import { VALIDATE } from '@/constants/validate';
 import useUserRegistrationStore from '@/store/user.registration.store';
-import { useAuth } from '@/features/auth/hooks/useAuth';
 
 const JoinStep2Nickname = () => {
   const { errorMessage, userCheckedNickname } = useAuth();
