+++ frontend/src/pages/Mypage/LikedReviews.tsx
@@ -1,7 +1,5 @@
 import Layout from '@/layout/Layout';
-import { ReviewList } from '@/components/Review';
-import { useLikedReviews } from '@/hooks/useInfiniteReviews';
-import { formatReviews } from '@/features/reviews/formatReviews';
+import { ReviewList, formatReviews, useLikedReviews } from '@/features/reviews';
 
 const LikedReviews = () => {
   const { data, isLoading, fetchNextPage, hasNextPage } = useLikedReviews();
